Mount the list and calendar views through one container each
RecordIndexListContainer and RecordListWidget both resolved the object
and its permissions, opened a RecordListContextProvider and rendered
RecordList with its SSE subscription; RecordIndexCalendarContainer and
RecordCalendarWidget did the same for the calendar and its three
effects. That wiring now lives once, in RecordListContainer and
RecordCalendarContainer, which take objectNameSingular and
viewBarInstanceId the way RecordBoardContainer does. The index page
mounts the list container directly, as it already did for the board.
The calendar keeps its index adapter because it guards on the current
view's calendarFieldMetadataId, while the widget guards on the
per-widget atom hydrated from a draft view and wraps a read-only
effect; both guards are kept as they were.

The RecordComponentInstanceContextsWrapper the two index containers
opened re-provided the six instance contexts RecordIndexContainerGater
already provides with the same recordIndexId, so it is dropped rather
than moved.
