+++ packages/twenty-front/src/modules/object-record/record-calendar-widget/components/RecordCalendarWidget.tsx
@@ -1,13 +1,7 @@
-import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useObjectPermissionsForObject } from '@/object-record/hooks/useObjectPermissionsForObject';
-import { RecordCalendar } from '@/object-record/record-calendar/components/RecordCalendar';
-import { RecordCalendarSSESubscribeEffect } from '@/object-record/record-calendar/components/RecordCalendarSSESubscribeEffect';
-import { RecordIndexCalendarDataLoaderEffect } from '@/object-record/record-calendar/components/RecordIndexCalendarDataLoaderEffect';
-import { RecordIndexCalendarSelectedDateInitEffect } from '@/object-record/record-calendar/components/RecordIndexCalendarSelectedDateInitEffect';
-import { RecordCalendarContextProvider } from '@/object-record/record-calendar/contexts/RecordCalendarContext';
+import { RecordCalendarContainer } from '@/object-record/record-calendar/components/RecordCalendarContainer';
 import { RecordCalendarWidgetReadOnlyEffect } from '@/object-record/record-calendar-widget/components/RecordCalendarWidgetReadOnlyEffect';
-import { recordIndexCalendarFieldMetadataIdComponentState } from '@/object-record/record-index/states/recordIndexCalendarFieldMetadataIdComponentState';
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
+import { recordIndexCalendarFieldMetadataIdComponentState } from '@/object-record/record-index/states/recordIndexCalendarFieldMetadataIdComponentState';
 import { useAtomComponentStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomComponentStateValue';
 import { styled } from '@linaria/react';
 import { isDefined } from 'twenty-shared/utils';
@@ -30,14 +24,6 @@ export const RecordCalendarWidget = ({
   const { objectNameSingular, recordIndexId, viewBarInstanceId } =
     useRecordIndexContextOrThrow();
 
-  const { objectMetadataItem } = useObjectMetadataItem({
-    objectNameSingular,
-  });
-
-  const objectPermissions = useObjectPermissionsForObject(
-    objectMetadataItem.id,
-  );
-
   // Hydrated per widget instance from the backing view (draft or persisted)
   // by the widget view load effect, so edit-mode previews work before save.
   const recordIndexCalendarFieldMetadataId = useAtomComponentStateValue(
@@ -56,20 +42,10 @@ export const RecordCalendarWidget = ({
         isReadOnly={isReadOnly}
       />
       <StyledCalendarContainer>
-        <RecordCalendarContextProvider
-          value={{
-            viewBarInstanceId,
-            objectNameSingular,
-            visibleRecordFields: [],
-            objectMetadataItem,
-            objectPermissions,
-          }}
-        >
-          <RecordCalendar />
-          <RecordCalendarSSESubscribeEffect />
-          <RecordIndexCalendarDataLoaderEffect />
-          <RecordIndexCalendarSelectedDateInitEffect />
-        </RecordCalendarContextProvider>
+        <RecordCalendarContainer
+          objectNameSingular={objectNameSingular}
+          viewBarInstanceId={viewBarInstanceId}
+        />
       </StyledCalendarContainer>
     </>
   );
