+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordIndexCalendarContainer.tsx
@@ -1,34 +1,11 @@
-import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { RecordComponentInstanceContextsWrapper } from '@/object-record/components/RecordComponentInstanceContextsWrapper';
-import { useObjectPermissionsForObject } from '@/object-record/hooks/useObjectPermissionsForObject';
-import { RecordCalendar } from '@/object-record/record-calendar/components/RecordCalendar';
-import { RecordCalendarSSESubscribeEffect } from '@/object-record/record-calendar/components/RecordCalendarSSESubscribeEffect';
-import { RecordIndexCalendarDataLoaderEffect } from '@/object-record/record-calendar/components/RecordIndexCalendarDataLoaderEffect';
-import { RecordIndexCalendarSelectedDateInitEffect } from '@/object-record/record-calendar/components/RecordIndexCalendarSelectedDateInitEffect';
-import { RecordCalendarContextProvider } from '@/object-record/record-calendar/contexts/RecordCalendarContext';
-
+import { RecordCalendarContainer } from '@/object-record/record-calendar/components/RecordCalendarContainer';
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
 import { useGetCurrentViewOnly } from '@/views/hooks/useGetCurrentViewOnly';
 import { isDefined } from 'twenty-shared/utils';
 
-type RecordIndexCalendarContainerProps = {
-  recordCalendarInstanceId: string;
-  viewBarInstanceId: string;
-};
-
-export const RecordIndexCalendarContainer = ({
-  viewBarInstanceId,
-  recordCalendarInstanceId,
-}: RecordIndexCalendarContainerProps) => {
-  const { objectNameSingular } = useRecordIndexContextOrThrow();
-
-  const { objectMetadataItem } = useObjectMetadataItem({
-    objectNameSingular,
-  });
-
-  const objectPermissions = useObjectPermissionsForObject(
-    objectMetadataItem.id,
-  );
+export const RecordIndexCalendarContainer = () => {
+  const { objectNameSingular, viewBarInstanceId } =
+    useRecordIndexContextOrThrow();
 
   const { currentView } = useGetCurrentViewOnly();
 
@@ -40,23 +17,9 @@ export const RecordIndexCalendarContainer = ({
   }
 
   return (
-    <RecordComponentInstanceContextsWrapper
-      componentInstanceId={recordCalendarInstanceId}
-    >
-      <RecordCalendarContextProvider
-        value={{
-          viewBarInstanceId,
-          objectNameSingular,
-          visibleRecordFields: [],
-          objectMetadataItem,
-          objectPermissions,
-        }}
-      >
-        <RecordCalendar />
-        <RecordCalendarSSESubscribeEffect />
-        <RecordIndexCalendarDataLoaderEffect />
-        <RecordIndexCalendarSelectedDateInitEffect />
-      </RecordCalendarContextProvider>
-    </RecordComponentInstanceContextsWrapper>
+    <RecordCalendarContainer
+      objectNameSingular={objectNameSingular}
+      viewBarInstanceId={viewBarInstanceId}
+    />
   );
 };
