+++ packages/twenty-front/src/modules/object-record/record-index/components/RecordIndexContainer.tsx
@@ -8,10 +8,10 @@ import { useAtomComponentStateValue } from '@/ui/utilities/state/jotai/hooks/use
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
 
 import { RecordIndexCalendarContainer } from '@/object-record/record-index/components/RecordIndexCalendarContainer';
-import { RecordIndexListContainer } from '@/object-record/record-index/components/RecordIndexListContainer';
 import { RecordIndexEmptyStateNotShared } from '@/object-record/record-index/components/RecordIndexEmptyStateNotShared';
 import { RecordIndexFiltersToContextStoreEffect } from '@/object-record/record-index/components/RecordIndexFiltersToContextStoreEffect';
 import { useHasCurrentViewNonReadableFields } from '@/object-record/record-index/hooks/useHasCurrentViewNonReadableFields';
+import { RecordListContainer } from '@/object-record/record-list/components/RecordListContainer';
 import { ViewType } from '@/views/types/ViewType';
 import { themeCssVariables } from 'twenty-ui/theme-constants';
 
@@ -64,16 +64,13 @@ export const RecordIndexContainer = () => {
           )}
           {recordIndexViewType === ViewType.CALENDAR && (
             <StyledContainerWithPadding>
-              <RecordIndexCalendarContainer
-                recordCalendarInstanceId={recordIndexId}
-                viewBarInstanceId={recordIndexId}
-              />
+              <RecordIndexCalendarContainer />
             </StyledContainerWithPadding>
           )}
           {recordIndexViewType === ViewType.LIST && (
             <StyledContainerWithPadding>
-              <RecordIndexListContainer
-                recordListInstanceId={recordIndexId}
+              <RecordListContainer
+                objectNameSingular={objectNameSingular}
                 viewBarInstanceId={recordIndexId}
               />
             </StyledContainerWithPadding>
