+++ packages/twenty-front/src/modules/object-record/record-list-widget/components/RecordListWidget.tsx
@@ -1,9 +1,5 @@
-import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
-import { useObjectPermissionsForObject } from '@/object-record/hooks/useObjectPermissionsForObject';
 import { useRecordIndexContextOrThrow } from '@/object-record/record-index/contexts/RecordIndexContext';
-import { RecordList } from '@/object-record/record-list/components/RecordList';
-import { RecordListSSESubscribeEffect } from '@/object-record/record-list/components/RecordListSSESubscribeEffect';
-import { RecordListContextProvider } from '@/object-record/record-list/contexts/RecordListContext';
+import { RecordListContainer } from '@/object-record/record-list/components/RecordListContainer';
 import { styled } from '@linaria/react';
 
 const StyledListContainer = styled.div`
@@ -18,27 +14,12 @@ export const RecordListWidget = () => {
   const { objectNameSingular, viewBarInstanceId } =
     useRecordIndexContextOrThrow();
 
-  const { objectMetadataItem } = useObjectMetadataItem({
-    objectNameSingular,
-  });
-
-  const objectPermissions = useObjectPermissionsForObject(
-    objectMetadataItem.id,
-  );
-
   return (
     <StyledListContainer>
-      <RecordListContextProvider
-        value={{
-          viewBarInstanceId,
-          objectNameSingular,
-          objectMetadataItem,
-          objectPermissions,
-        }}
-      >
-        <RecordList />
-        <RecordListSSESubscribeEffect />
-      </RecordListContextProvider>
+      <RecordListContainer
+        objectNameSingular={objectNameSingular}
+        viewBarInstanceId={viewBarInstanceId}
+      />
     </StyledListContainer>
   );
 };
